Improve user management and token check
