+++ src/utils/handlers/authenticationHandlers.js
@@ -1,6 +1,7 @@
 import axios from 'axios';
 import { pathObj, tokenKey } from '../generalVariables';
 import { setHeaders } from '../requestHeaders';
+import * as jwt_decode from 'jwt-decode';
 
 export const loginHandler = ({ username, password }) => {
   // With the finalization of the database schema, more checks can be implemented (with separate error-messages)
@@ -85,6 +86,20 @@ export const checkAndStoreToken = token => {
 export const logout = () => {
   if (localStorage.getItem(tokenKey)) {
     localStorage.removeItem(tokenKey);
+  }
+};
+
+export const getUser = () => {
+  const token = localStorage.getItem(tokenKey);
+  if (token) {
+    const decodedToken = jwt_decode(token);
+    if (decodedToken.exp * 1000 < Date.now()) {
+      localStorage.removeItem(tokenKey);
+      return false;
+    }
+    return decodedToken;
+  } else {
+    return false;
   }
 };
